SimpleWebScraper/SimpleWebScraper/Workers/Scraper.cs
using SimpleWebScraper.Data;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using System.Text.RegularExpressions;

namespace SimpleWebScraper.Workers
{
    class Scraper
    {
        public List<string> Scrape(ScrapeCriteria scrapeCriteria)
        {
            List<string> scrapedElements = new List<string>(); // this is what we return 
            //begin matching
            MatchCollection matches = Regex.Matches(scrapeCriteria.Data, scrapeCriteria.Regex, scrapeCriteria.RegexOption);

            foreach (Match match in matches)  
            {
                if (!scrapeCriteria.Parts.Any()) // if we dont need to go deeper
                {
                    scrapedElements.Add(match.Groups[0].Value); // 
                }
                else // going deeper and grabbing certain parts of element
                {
                    foreach (var part in scrapeCriteria.Parts)  
                    {
                        Match matchedPart = Regex.Match(match.Groups[0].Value, part.Regex, part.RegexOption);

                        if (matchedPart.Success) scrapedElements.Add(matchedPart.Groups[1].Value);
                    }
                }
            }

            return scrapedElements;
        }
    }
}
